dev-requirements.txt
-r requirements.txt
flake8
discord.py-stubs
